+++ server/router/routes/menu.js
@@ -1,9 +1,9 @@
 'use strict'
 
 import Menu from '../../model/menu'
-// import _ from 'lodash'
 import { JsonError } from '../../error'
 import { isBearerAuthenticated } from '../../auth'
+import { hasPermission } from '../../util'
 
 /**
  * 判断当前菜单是否有权限
@@ -11,48 +11,33 @@ import { isBearerAuthenticated } from '../../auth'
  * @param  {[type]}  permissions [description]
  * @return {Boolean}             [description]
  */
-function isMenuHasPermission(menu, permissions){
+function isMenuShowHasPermission(menu, permissions){
 	//判断是否为根目录,根目录所有用户都有权限
 	if(!menu.parent){
 		return true;
 	}
-	for (var i = 0; i < permissions.length; i++) {
-		if(menu.shortname == permissions[i].name && permissions[i].permissions.indexOf('查看') != -1){
-			return true;
-		}
-	}	
-	return false;
+	//判断是否有查看权限
+	return hasPermission(permissions, menu.shortname, '查看');
 }
 
 
 export default (router => {
 	router
-		// .get('/menus/tree', async (ctx) =>{
-		// 	try{
-		// 		//查询所有菜单
-		// 		const menus = await Menu.find({parent: {$exists: false}}).populate('children').sort('order').exec();
-		// 		ctx.body = {
-		// 			code: 10000,
-		// 			result: menus
-		// 		}
-		// 	}catch(error){
-		// 		console.log(error);
-		// 		throw new JsonError(20005, '树级菜单获取失败')
-		// 	}
-		// })
 		//根据用户权限返回相关菜单
 		.get('/menus/listWithPermission', isBearerAuthenticated(), async (ctx) =>{
 			try{
+				const account = ctx.state.user;
 				//获取用户权限
-				const permissions = ctx.state.user.permissions;
-				// console.log(permissions);
+				const permissions = account.permissions;
 				//查询所有菜单
 				const menus_ = await Menu.find();
 				let menus = [];
-				for (let i = 0; i <menus_.length; i++) {
-					const isHas = isMenuHasPermission(menus_[i], permissions);
-					if(isHas){
-						menus.push(menus_[i]);
+				//处理菜单'查看'权限
+				for (let i = 0; i < menus_.length; i++) {
+					const m = menus_[i];
+					const isHasPermission = isMenuShowHasPermission(m, permissions);
+					if(isHasPermission){
+						menus.push(m);
 					}
 				}
 				ctx.body = {
@@ -61,12 +46,19 @@ export default (router => {
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(20005, '菜单获取失败');
 			}
 		})
-		.get('/menus/list', async (ctx) => {
+		.get('/menus/list',isBearerAuthenticated(), async (ctx) => {
 			try{
-				
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'menu', '查看')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const menus = await Menu.find();
 				ctx.body = {
 					code: 10000,
@@ -74,11 +66,19 @@ export default (router => {
 				};
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(20001, '菜单获取失败');
 			}
 		})
-		.put('/menus/add', async (ctx) => {
+		.put('/menus/add',isBearerAuthenticated(), async (ctx) => {
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'menu', '创建')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const {name, shortname, path, componentUrl, icon, order, isHidden, parent} = ctx.request.body;
 				await Menu.create({name,shortname, path, componentUrl, icon, order, isHidden, parent});
 				ctx.body = {
@@ -87,11 +87,19 @@ export default (router => {
 				};
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(20002, '菜单创建失败');
 			}
 		})
-		.post('/menus/edit', async (ctx) => {
+		.post('/menus/edit',isBearerAuthenticated(), async (ctx) => {
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'menu', '编辑')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const {_id, name,shortname, path, componentUrl, icon, order, isHidden, parent} = ctx.request.body;
 				await Menu.findByIdAndUpdate(_id, { name,shortname, path, componentUrl, icon, order, isHidden, parent })
 				ctx.body = {
@@ -100,11 +108,19 @@ export default (router => {
 				};
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(20003, '菜单更新失败');
 			}
 		})
-		.delete('/menus/remove/:id', async (ctx) =>{
+		.delete('/menus/remove/:id',isBearerAuthenticated(), async (ctx) =>{
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'menu', '删除')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const _id = ctx.params.id;
 				await Menu.findByIdAndRemove(_id);
 				ctx.body = {
@@ -113,6 +129,9 @@ export default (router => {
 				};
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(20004, '菜单删除失败');
 			}
