+++ server/router/routes/permission.js
@@ -3,22 +3,17 @@
 import Permission from '../../model/permission'
 import { JsonError } from '../../error'
 import { isBearerAuthenticated } from '../../auth'
+import { hasPermission } from '../../util'
 
 export default (router=>{
 	router
-		// .post('/permission/user',isBearerAuthenticated(), async (ctx)=>{
-		// 	try{
-		// 		console.log(ctx.state.user);
-		// 		ctx.body = {
-		// 			code: 10000
-		// 		}
-		// 	}catch(error){
-		// 		console.log(error)
-		// 		throw new JsonError(50005, '获取用户权限失败')
-		// 	}
-		// })
-		.post('/permission/list', async (ctx)=>{
+		.post('/permission/list',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permission = ctx.state.user.permissions;
+				if(!hasPermission(permission, 'permission', '查看')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const {name} = ctx.request.body;
 				let query = {};
 				if(name){
@@ -33,11 +28,19 @@ export default (router=>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(50001, '权限获取失败')
 			}
 		})
-		.put('/permission/add', async (ctx)=>{
+		.put('/permission/add',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permission = ctx.state.user.permissions;
+				if(!hasPermission(permission, 'permission', '创建')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const {name, order, role, permissions} = ctx.request.body;
 				await Permission.create({
 					name, order, role ,permissions
@@ -48,11 +51,19 @@ export default (router=>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(50002, '权限创建失败')
 			}
 		})
-		.post('/permission/edit', async (ctx)=>{
+		.post('/permission/edit',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permission = ctx.state.user.permissions;
+				if(!hasPermission(permission, 'permission', '编辑')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const {_id, name, order, role, permissions} = ctx.request.body;
 				await Permission.findByIdAndUpdate(_id, {
 					name, order, role, permissions
@@ -63,11 +74,19 @@ export default (router=>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(50003, '权限更新失败')
 			}
 		})
-		.delete('/permission/remove/:id', async (ctx)=>{
+		.delete('/permission/remove/:id',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permission = ctx.state.user.permissions;
+				if(!hasPermission(permission, 'permission', '删除')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const _id = ctx.params.id;
 				await Permission.findByIdAndRemove(_id);
 				ctx.body = {
@@ -76,6 +95,9 @@ export default (router=>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(50004, '权限删除失败');
 			}
 		})
