+++ server/router/routes/role.js
@@ -2,11 +2,18 @@
 
 import Role from '../../model/role'
 import { JsonError } from '../../error'
+import { isBearerAuthenticated } from '../../auth'
+import { hasPermission } from '../../util'
 
 export default (router =>{
 	router
-		.post('/role/list', async (ctx)=>{
+		.post('/role/list',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'role', '查看')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const { name } = ctx.request.body;
 				let query = {};
 				//判断是否有姓名,如果有就是从查询过来
@@ -22,11 +29,19 @@ export default (router =>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(40001, '角色获取失败');
 			}
 		})
-		.put('/role/add', async (ctx)=>{
+		.put('/role/add',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'role', '创建')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const { name, order, accounts} = ctx.request.body;
 				await Role.create({
 					name, order, accounts
@@ -37,11 +52,19 @@ export default (router =>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(40002, '角色创建失败');
 			}
 		})
-		.post('/role/edit', async (ctx)=>{
+		.post('/role/edit',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'role', '编辑')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const {_id, name, order, accounts} = ctx.request.body;
 				await Role.findByIdAndUpdate(_id, {
 					name, order, accounts
@@ -52,11 +75,19 @@ export default (router =>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(40003, '角色更新失败');
 			}
 		})
-		.delete('/role/remove/:id', async (ctx)=>{
+		.delete('/role/remove/:id',isBearerAuthenticated(), async (ctx)=>{
 			try{
+				const permissions = ctx.state.user.permissions;
+				if(!hasPermission(permissions, 'role', '删除')){
+					throw new JsonError(20009, '您无操作权限，请联系管理员');
+					return;
+				}
 				const _id = ctx.params.id;
 				await Role.findByIdAndRemove(_id);
 				ctx.body = {
@@ -65,6 +96,9 @@ export default (router =>{
 				}
 			}catch(error){
 				console.log(error);
+				if(error instanceof JsonError){
+					throw error;
+				}
 				throw new JsonError(40004, '角色删除失败');
 			}
 		})
